change cron job to single event that backfills
