+++ src/app/api/inngest/route.ts
@@ -1,12 +1,12 @@
 import { serve } from 'inngest/next'
 
 import { airdropNFTWithInngest } from '@/inngest/functions/airdropNFT'
-import {
-  auditUsersTotalDonationAmountUsdInngestCronJob,
-  auditUsersTotalDonationAmountUsdInngestCronJobUpdateBatchOfUsers,
-} from '@/inngest/functions/auditUsersTotalDonationAmountUsdCronJob'
 import { backfillNFTWithInngest } from '@/inngest/functions/backfillNFT'
 import { backfillNFTInngestCronJob } from '@/inngest/functions/backfillNFTCronJob'
+import {
+  backfillUsersTotalDonationAmountUsdInngest,
+  backfillUsersTotalDonationAmountUsdInngestUpdateBatchOfUsers,
+} from '@/inngest/functions/backfillUsersTotalDonationAmountUsd'
 import { cleanupNFTMintsWithInngest } from '@/inngest/functions/cleanupNFTMints'
 import { cleanupPostalCodesWithInngest } from '@/inngest/functions/cleanupPostalCodes'
 import { emailRepViaCapitolCanaryWithInngest } from '@/inngest/functions/emailRepViaCapitolCanary'
@@ -29,7 +29,7 @@ export const { GET, POST, PUT } = serve({
     backfillNFTWithInngest,
     backfillNFTInngestCronJob,
     cleanupNFTMintsWithInngest,
-    auditUsersTotalDonationAmountUsdInngestCronJob,
-    auditUsersTotalDonationAmountUsdInngestCronJobUpdateBatchOfUsers,
+    backfillUsersTotalDonationAmountUsdInngest,
+    backfillUsersTotalDonationAmountUsdInngestUpdateBatchOfUsers,
   ],
 })
